Extend Docker entrypoint.sh to support installing extra-system-packages
The docker container currently contains the bare minimum list of packages
to get an image built. By adding a parameter to the entrypoint.sh script
more packages can be installed along the way in a flexible way, without
the need to extend the provided container.

E.g. useable to add bmap-tools or zstd.