Disable packer qemu setup inside Docker and switch to tonistiigi/binfmt
tonistiigi/binfmt is a drop in replacement for multiarch/qemu-user-static,
which not only can emulate arm/arm64 on amd64 but also on arm64 and other
architectures. By setting up qemu with Docker can automatically execute
the binaries of other achitecture in the chroot without any additional
setup.

This way the Docker Container also works on Linux aarch64 including Docker
on Apple M1 hardware.