Handle chroot mounts with symlink as mountpoint
During cleanup of the `step_setup_chroot`, we will try to unmount
all paths mounted by step's Run. This is done be reading mtab
and looking if each of our mountpoints is present there.

Unfortunately, when mountpoint is a symlink to other directory,
mtab will contain the real canonical path. In such cases we would
skip unmounting directory and later (usually) fail to unmount
device partitions. Moreover, we will endup with still attached
loop device because `losetup -d` will only schedule detachment
once device is no longer needed.

To avoid this fiasco, we try to resolve paths to canonical path
when unmounting. If this succeed, we will use that path to unmount,
if not normal logic applies. Whenever a mountpoint is considered
as already unmounted (for example by the user provisioning script),
that fact is now logged.

Additionally, some logging messages were added (visible with
`PACKER_LOG`) to help debug the errors of unmounting and busy
file systems. Unfortunately regular umount error does not contain
the path and might be mistaken by 'optional' fuser failure.