Use losetup find and attach to avoid race condition
It may happen that between checking for free loopback device
and setting it up, another process will take that device.

There is a high chance for collision when multiple parallel builds
are being done from the same image file as seen in #163.

Instead of manually trying to find and acquire device, lets
use the losetup for that. When looking for free device, according
to manual [1] we can also specify a image to attach.

To know the device which is mapped to image, additional flag
is set, so that losetup will print the mapped device.

Short flags are replaced with longer version for easier
understanding of their meaning.

[1]: https://man.archlinux.org/man/losetup.8.en