Add repartition method for modifying existing partitions
Currently, the 'new' method can be used to effectively manage partitions
in an image when the source image does not have the desired layout.
However, it relies on a root filesystem archive, which is not available
in all cases (e.g. Ubuntu does not publish Raspberry Pi rootfs
archives).

The 'repartition' method allows for taking an existing image and
changing its partitions. That is, partitions may be arbitrarily created,
grown, shrunk, or destroyed: this is done according to the partition
layout we give in the image config.

This adds two new options to partition configuration:

* `resize_fs`: Whether to run resize2fs on the partition after
  repartitioning. This is necessary because:
  - Not all partitions can be resized with resize2fs (e.g. NTFS)
  - Only the user knows whether or not a partition from the source image
    has been resized
* `skip_mkfs`: If true, the partition will **not** be formatted with
  mkfs in repartition mode.